+++ src/components/KakaoMap.tsx
@@ -1,6 +1,9 @@
 import React, { useEffect } from "react";
 import styled from "styled-components";
-import { defaultLatitude, defaultLongitude } from "../constants/commonCoordinates";
+import {
+  defaultLatitude,
+  defaultLongitude,
+} from "../constants/commonCoordinates";
 
 declare global {
   interface Window {
@@ -9,7 +12,6 @@ declare global {
 }
 
 const KakaoMap: React.FC = () => {
-
   useEffect(() => {
     const container = document.getElementById("map");
     const markerPosition = new window.kakao.maps.LatLng(
